Remove microformats gem dependency (#22923)
Looks like this gem was introduced as a dependency in 89707ad0ac for
testing Miroformat output.  The last test using the Microformats gem was
removed in 62782babd08bc2385a604e275bf88af925d137c1, so I think it is
safe to remove this dependency.

For context, you [can't install the microformats gem with Ruby 3.2](https://github.com/microformats/microformats-ruby/pull/131),
so we can't currently bundle Mastodon with Ruby 3.2.  But since we don't
really need this gem, we can just remove it and unblock Ruby 3.2